Replace cryptoswift with cryptoswiftwrapper
A Swift package wrapper of CryptoSwift in binary form (via an XCFramework bundle).
The wrapper is necessary for multiple targets to depend on the binary through
Swift Package Manager, as only one binary target can be defined for it within a dependency graph.
The release versions of this package mirrors the ones of CryptoSwiftstarting from 1.4.3.